unified-diet-tracker/backend/requirements.txt
flask
requests
colorama
pandas
python-docx
flask-sqlalchemy
psycopg2-binary
python-dotenv
gunicorn
flask-cors
ollama
